04_apis_gpt_actions/13_oauth2/04_authentication/requirements.txt
fastapi
uvicorn[standard]
pydantic
SQLAlchemy
python-multipart
psycopg2-binary
email-validator

python-dotenv
python-jose
passlib

google-api-python-client  
google-auth 
google-auth-oauthlib 
google-auth-httplib2

itsdangerous